#9 Move reconciler to adapters
