+++ pkg/adapter/reconciler/interfaces.go
