+++ pkg/adapter/reconciler/mock_deleteArchiveHandler_test.go
